Moved Containers into Components Folder
The containers are just special components. They handle their own data
dependencies. The containers aren't special enough to justify being put
in their own folder. They are just components which wrap other
components.

See #32
