+++ src/components/ArtistContainer/component/style.css
