+++ src/components/ArtistContainer/index.tsx
