+++ src/components/Footer/NowPlaying.tsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom';
 import Artwork from '../Artwork';
 import { join, unique } from '../../utils';
 import { album, artist } from '../../utils/paths';
-import { Song } from '../../containers/Footer/enhancers/query';
+import { Song } from '../FooterContainer/enhancers/query';
 import styles from './NowPlaying.css';
 import DefaultCover from '../icons/DefaultCover';
