+++ src/components/QueueContainer/component/styles.css
